Extract common logic to the composable.
